- added a cache for directories offspring count.
- added unix name in struct dir.

- added pread/pwrite, (there was at least one bug in adf_off book keeping).

- removed (temporary?) memory mmap header, it was unused and made the code
  unreadable.

- used a new struct path with unix name, mac name and file stat.

- fixed for catsearch (not well tested)

- temporary ATACC and NO_UAM_LOAD, I will remove them (ATACC is a test coverage tool).

- added AFP 3.1 calls readext,writeext, bytelockext, enumerateext2
  (catsearchext, enumerateext are the same) . --afp3 parameter in
  configure. configure.in changes aren't good. Maybe get CVS samba stuff
  (eg for linux there's no sendfile64 on 32 bits box). The biggest change
  is that we can't use the ressource fork len in the header (it's too
  small). Not well tested.

- fixed few signed/unsigned missmatch, mainly when a signed fetched from the wire is checked
  against and unsigned buffer size.

+++ libatalk/dsi/dsi_stream.c
@@ -1,5 +1,5 @@
 /*
- * $Id: dsi_stream.c,v 1.8 2002-02-02 21:09:13 jmarcus Exp $
+ * $Id: dsi_stream.c,v 1.9 2002-10-11 14:18:39 didg Exp $
  *
  * Copyright (c) 1998 Adrian Sun (asun@zoology.washington.edu)
  * All rights reserved. See COPYRIGHT.
@@ -96,7 +96,7 @@ int dsi_stream_send(DSI *dsi, void *buf, size_t length)
   sigset_t oldset;
 #ifdef USE_WRITEV
   struct iovec iov[2];
-  size_t  towrite;
+  size_t towrite;
   ssize_t len;
 #endif /* USE_WRITEV */
 
@@ -170,8 +170,8 @@ int dsi_stream_send(DSI *dsi, void *buf, size_t length)
 /* read data. function on success. 0 on failure. data length gets
  * stored in length variable. this should really use size_t's, but
  * that would require changes elsewhere. */
-int dsi_stream_receive(DSI *dsi, void *buf, const int ilength,
-		       int *rlength)
+int dsi_stream_receive(DSI *dsi, void *buf, const size_t ilength,
+		       size_t *rlength)
 {
   char block[DSI_BLOCKSIZ];
